DIGITAL-670: Restore build-and-deploy, with draft CMS restrictions
Changes to the build-and-deploy workflow include how to get the nginx buildpack version
and extract the latest version of nginx therein, both due to changes in cloud foundry.

Also backed out tf code that attempted to define a set of allowed IP addresses for the CMS,
but introduced a cycle. Replaced this with a simpler set of allowed addresses and corrected
the order of directives in nginx configuration.